src/config.py
import os
from dotenv import load_dotenv

load_dotenv()

MLFLOW_TRACKING_URI = os.getenv("MLFLOW_TRACKING_URI", "http://localhost:5000")
MLFLOW_EXPERIMENT_NAME = os.getenv("MLFLOW_EXPERIMENT_NAME", "breast_cancer_mlops")
MODEL_PATH = os.getenv("MODEL_PATH", "models/exported/model.pkl")
MODEL_VERSION = os.getenv("MODEL_VERSION", "v1")

RANDOM_STATE = int(os.getenv("RANDOM_STATE", "42"))
TEST_SIZE = float(os.getenv("TEST_SIZE", "0.2"))
