requirements.txt
scikit-learn
pandas
numpy
matplotlib
seaborn
dvc

mlflow<3.0
optuna
zenml

fastapi
uvicorn
pydantic

joblib
python-dotenv

boto3
minio
sqlalchemy-utils
sqlmodel
passlib
pymysql
bcrypt